+++ Changelog.md
@@ -2,6 +2,8 @@
 
 ## master (unreleased)
 
+- Fix warning in China's holidays to dynamically read supported years.
+
 Nothing here yet.
 
 ## v8.0.2 (2020-01-24)